// src/modules/notifications/enums/notification-event-type.enum.ts
export enum NotificationEventType {
  BOOKING_CREATED = 'BOOKING_CREATED',
  BOOKING_ACCEPTED = 'BOOKING_ACCEPTED',
  BOOKING_REJECTED = 'BOOKING_REJECTED',
  QUOTE_CREATED = 'QUOTE_CREATED',
  PAYMENT_SUCCEEDED = 'PAYMENT_SUCCEEDED',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  TRUCK_POST_PUBLISHED = 'TRUCK_POST_PUBLISHED',
  REWARD_EARNED = 'REWARD_EARNED',
  CHECK_IN_VERIFIED = 'CHECK_IN_VERIFIED',
  VENDOR_VERIFICATION_SUBMITTED = 'VENDOR_VERIFICATION_SUBMITTED',
  BOOKING_ISSUE_REPORTED = 'BOOKING_ISSUE_REPORTED',
  BOOKING_ISSUE_MESSAGE_CREATED = 'BOOKING_ISSUE_MESSAGE_CREATED',
  PAYMENT_ATTENTION_REQUIRED = 'PAYMENT_ATTENTION_REQUIRED',
  CONTENT_REPORT_SUBMITTED = 'CONTENT_REPORT_SUBMITTED',
}